refactor: reorganize PluginsClient files in langchain format
