+++ api/app/clients/llm/index.js
@@ -1,0 +1,5 @@
+const createLLM = require('./createLLM');
+
+module.exports = {
+  createLLM,
+};
