+++ api/app/clients/output_parsers/handleOutputs.js
@@ -1,8 +1,4 @@
-const {
-  instructions,
-  imageInstructions,
-  errorInstructions,
-} = require('../../prompts/instructions');
+const { instructions, imageInstructions, errorInstructions } = require('../prompts/instructions');
 
 function getActions(actions = [], functionsAgent = false) {
   let output = 'Internal thoughts & actions taken:\n"';
