+++ api/app/clients/output_parsers/index.js
@@ -1,9 +1,7 @@
 const addImages = require('./addImages');
-const createLLM = require('./createLLM');
 const handleOutputs = require('./handleOutputs');
 
 module.exports = {
   addImages,
-  createLLM,
   ...handleOutputs,
 };
